+++ client/client_integration_common_test.go
@@ -62,7 +62,7 @@ func TestQuerySetStatements(t *testing.T) {
 		context.TODO(),
 		engineUrlMock,
 		query,
-		map[string]string{"timezone": "America/New_York", "database": databaseMock},
+		map[string]string{"time_zone": "America/New_York", "database": databaseMock},
 		ConnectionControl{},
 	)
 	if err != nil {
