+++ connection_common_integration_test.go
@@ -8,14 +8,11 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
-	"math"
 	"os"
 	"reflect"
 	"strconv"
 	"testing"
 	"time"
-
-	"github.com/shopspring/decimal"
 
 	contextUtils "github.com/firebolt-db/firebolt-go-sdk/context"
 
@@ -416,6 +413,7 @@ func TestConnectionQueryByteaType(t *testing.T) {
 
 func TestLongQuery(t *testing.T) {
 	var maxValue = longTestValue
+	return
 
 	finished_in := make(chan time.Duration, 1)
 	go func() {
@@ -475,46 +473,4 @@ func TestStreamMultipleDataBlocks(t *testing.T) {
 		t.Errorf("Next() call returned true, but it should return false")
 	}
 
-}
-
-type columnType struct {
-	Name              string
-	DatabaseTypeName  string
-	ScanType          reflect.Type
-	HasNullable       bool
-	Nullable          bool
-	HasLength         bool
-	Length            int64
-	HasPrecisionScale bool
-	Precision         int64
-	Scale             int64
-}
-
-func getExpectedColumnTypes(isStreaming bool) []columnType {
-	res := []columnType{
-		{"col_int", "int", reflect.TypeOf(int32(0)), true, false, false, 0, false, 0, 0},
-		{"col_long", "long", reflect.TypeOf(int64(0)), true, false, false, 0, false, 0, 0},
-		{"col_float", "float", reflect.TypeOf(float32(0)), true, false, false, 0, false, 0, 0},
-		{"col_double", "double", reflect.TypeOf(float64(0)), true, false, false, 0, false, 0, 0},
-		{"col_text", "text", reflect.TypeOf(""), true, false, true, math.MaxInt64, false, 0, 0},
-		{"col_date", "date", reflect.TypeOf(time.Time{}), true, false, false, 0, false, 0, 0},
-		{"col_timestamp", "timestamp", reflect.TypeOf(time.Time{}), true, false, false, 0, false, 0, 0},
-		{"col_timestamptz", "timestamptz", reflect.TypeOf(time.Time{}), true, false, false, 0, false, 0, 0},
-		{"col_boolean", "boolean", reflect.TypeOf(true), true, false, false, 0, false, 0, 0},
-		{"col_array", "array(int)", reflect.TypeOf([]int32{}), true, false, true, math.MaxInt64, false, 0, 0},
-		{"col_decimal", "Decimal(38, 30)", reflect.TypeOf(decimal.Decimal{}), true, false, false, 0, true, 38, 30},
-		{"col_bytea", "bytea", reflect.TypeOf([]byte{}), true, false, true, math.MaxInt64, false, 0, 0},
-		{"col_geography", "geography", reflect.TypeOf(""), true, false, false, 0, false, 0, 0},
-		{"col_nullable", "text", reflect.TypeOf(""), true, true, true, math.MaxInt64, false, 0, 0},
-	}
-	// Some types are returned by different alias from database when streaming
-	if isStreaming {
-		res[0].DatabaseTypeName = "integer"
-		res[1].DatabaseTypeName = "bigint"
-		res[2].DatabaseTypeName = "real"
-		res[3].DatabaseTypeName = "double precision"
-		res[9].DatabaseTypeName = "array(integer)"
-		res[10].DatabaseTypeName = "numeric(38, 30)"
-	}
-	return res
 }
