+++ connection_integration_v0_test.go
@@ -8,11 +8,14 @@ import (
 	"database/sql"
 	"database/sql/driver"
 	"fmt"
+	"math"
+	"reflect"
 	"runtime/debug"
 	"testing"
 	"time"
 
-	contextUtils "github.com/firebolt-db/firebolt-go-sdk/context"
+	"github.com/shopspring/decimal"
+
 	"github.com/firebolt-db/firebolt-go-sdk/utils"
 )
 
@@ -112,6 +115,38 @@ func TestConnectionPreparedStatementV0(t *testing.T) {
 	}
 }
 
+type columnType struct {
+	Name              string
+	DatabaseTypeName  string
+	ScanType          reflect.Type
+	HasNullable       bool
+	Nullable          bool
+	HasLength         bool
+	Length            int64
+	HasPrecisionScale bool
+	Precision         int64
+	Scale             int64
+}
+
+func getExpectedColumnTypes() []columnType {
+	res := []columnType{
+		{"col_int", "int", reflect.TypeOf(int32(0)), true, false, false, 0, false, 0, 0},
+		{"col_long", "long", reflect.TypeOf(int64(0)), true, false, false, 0, false, 0, 0},
+		{"col_float", "float", reflect.TypeOf(float32(0)), true, false, false, 0, false, 0, 0},
+		{"col_double", "double", reflect.TypeOf(float64(0)), true, false, false, 0, false, 0, 0},
+		{"col_text", "text", reflect.TypeOf(""), true, false, true, math.MaxInt64, false, 0, 0},
+		{"col_date", "date", reflect.TypeOf(time.Time{}), true, false, false, 0, false, 0, 0},
+		{"col_timestamp", "timestamp", reflect.TypeOf(time.Time{}), true, false, false, 0, false, 0, 0},
+		{"col_timestamptz", "timestamptz", reflect.TypeOf(time.Time{}), true, false, false, 0, false, 0, 0},
+		{"col_boolean", "boolean", reflect.TypeOf(true), true, false, false, 0, false, 0, 0},
+		{"col_array", "array(int)", reflect.TypeOf([]int32{}), true, false, true, math.MaxInt64, false, 0, 0},
+		{"col_decimal", "Decimal(38, 30)", reflect.TypeOf(decimal.Decimal{}), true, false, false, 0, true, 38, 30},
+		{"col_bytea", "bytea", reflect.TypeOf([]byte{}), true, false, true, math.MaxInt64, false, 0, 0},
+		{"col_nullable", "text", reflect.TypeOf(""), true, true, true, math.MaxInt64, false, 0, 0},
+	}
+	return res
+}
+
 // TestResponseMetadata is the same as for V2 but without new types (like geography)
 func TestResponseMetadata(t *testing.T) {
 	const selectAllTypesSQL = `
@@ -129,47 +164,45 @@ func TestResponseMetadata(t *testing.T) {
        'abc123'::bytea                                           as col_bytea,
        null                                                      as col_nullable;`
 
-	utils.RunInMemoryAndStream(t, func(t *testing.T, ctx context.Context) {
-		expectedColumnTypes := getExpectedColumnTypes(contextUtils.IsStreaming(ctx))
+	ctx := context.Background()
+	expectedColumnTypes := getExpectedColumnTypes()
 
-		conn, err := sql.Open("firebolt", dsnMock)
-		if err != nil {
-			t.Errorf(OPEN_CONNECTION_ERROR_MSG)
-			t.FailNow()
-		}
+	conn, err := sql.Open("firebolt", dsnMock)
+	if err != nil {
+		t.Errorf(OPEN_CONNECTION_ERROR_MSG)
+		t.FailNow()
+	}
 
-		rows, err := conn.QueryContext(ctx, selectAllTypesSQL)
-		if err != nil {
-			t.Errorf(STATEMENT_ERROR_MSG, err)
-			t.FailNow()
-		}
+	rows, err := conn.QueryContext(ctx, selectAllTypesSQL)
+	if err != nil {
+		t.Errorf(STATEMENT_ERROR_MSG, err)
+		t.FailNow()
+	}
 
-		if !rows.Next() {
-			t.Errorf("Next() call returned false with error: %v", rows.Err())
-			t.FailNow()
-		}
+	if !rows.Next() {
+		t.Errorf("Next() call returned false with error: %v", rows.Err())
+		t.FailNow()
+	}
 
-		types, err := rows.ColumnTypes()
-		if err != nil {
-			t.Errorf("ColumnTypes returned an error, but shouldn't")
-			t.FailNow()
-		}
+	types, err := rows.ColumnTypes()
+	if err != nil {
+		t.Errorf("ColumnTypes returned an error, but shouldn't")
+		t.FailNow()
+	}
 
-		for i, ct := range types {
-			utils.AssertEqual(ct.Name(), expectedColumnTypes[i].Name, t, fmt.Sprintf("column name is not equal for column %s", ct.Name()))
-			utils.AssertEqual(ct.DatabaseTypeName(), expectedColumnTypes[i].DatabaseTypeName, t, fmt.Sprintf("database type name is not equal for column %s", ct.Name()))
-			utils.AssertEqual(ct.ScanType(), expectedColumnTypes[i].ScanType, t, fmt.Sprintf("scan type is not equal for column %s", ct.Name()))
-			nullable, ok := ct.Nullable()
-			utils.AssertEqual(ok, expectedColumnTypes[i].HasNullable, t, fmt.Sprintf("nullable ok is not equal for column %s", ct.Name()))
-			utils.AssertEqual(nullable, expectedColumnTypes[i].Nullable, t, fmt.Sprintf("nullable is not equal for column %s", ct.Name()))
-			length, ok := ct.Length()
-			utils.AssertEqual(ok, expectedColumnTypes[i].HasLength, t, fmt.Sprintf("length ok is not equal for column %s", ct.Name()))
-			utils.AssertEqual(length, expectedColumnTypes[i].Length, t, fmt.Sprintf("length is not equal for column %s", ct.Name()))
-			precision, scale, ok := ct.DecimalSize()
-			utils.AssertEqual(ok, expectedColumnTypes[i].HasPrecisionScale, t, fmt.Sprintf("precision scale ok is not equal for column %s", ct.Name()))
-			utils.AssertEqual(precision, expectedColumnTypes[i].Precision, t, fmt.Sprintf("precision is not equal for column %s", ct.Name()))
-			utils.AssertEqual(scale, expectedColumnTypes[i].Scale, t, fmt.Sprintf("scale is not equal for column %s", ct.Name()))
-		}
-
-	})
+	for i, ct := range types {
+		utils.AssertEqual(ct.Name(), expectedColumnTypes[i].Name, t, fmt.Sprintf("column name is not equal for column %s", ct.Name()))
+		utils.AssertEqual(ct.DatabaseTypeName(), expectedColumnTypes[i].DatabaseTypeName, t, fmt.Sprintf("database type name is not equal for column %s", ct.Name()))
+		utils.AssertEqual(ct.ScanType(), expectedColumnTypes[i].ScanType, t, fmt.Sprintf("scan type is not equal for column %s", ct.Name()))
+		nullable, ok := ct.Nullable()
+		utils.AssertEqual(ok, expectedColumnTypes[i].HasNullable, t, fmt.Sprintf("nullable ok is not equal for column %s", ct.Name()))
+		utils.AssertEqual(nullable, expectedColumnTypes[i].Nullable, t, fmt.Sprintf("nullable is not equal for column %s", ct.Name()))
+		length, ok := ct.Length()
+		utils.AssertEqual(ok, expectedColumnTypes[i].HasLength, t, fmt.Sprintf("length ok is not equal for column %s", ct.Name()))
+		utils.AssertEqual(length, expectedColumnTypes[i].Length, t, fmt.Sprintf("length is not equal for column %s", ct.Name()))
+		precision, scale, ok := ct.DecimalSize()
+		utils.AssertEqual(ok, expectedColumnTypes[i].HasPrecisionScale, t, fmt.Sprintf("precision scale ok is not equal for column %s", ct.Name()))
+		utils.AssertEqual(precision, expectedColumnTypes[i].Precision, t, fmt.Sprintf("precision is not equal for column %s", ct.Name()))
+		utils.AssertEqual(scale, expectedColumnTypes[i].Scale, t, fmt.Sprintf("scale is not equal for column %s", ct.Name()))
+	}
 }
